backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
TTS==0.14.3
numpy==1.24.3
torch==2.0.1